feat(tool-server): add search_shapes tool and extract search logic
- Extract shape search algorithms (soundex, token splitting, scored AND/OR ranking) from `mcp-app-server` into a new module `shared/shape-search.js` to prevent duplication.
- Update `mcp-app-server` to import shape search functions from the new shared location.
- Add `shared/shape-search.js` and `shape-search/search-index.json` to the `copy-shared` build step in `mcp-tool-server/package.json`.
- Implement `search_shapes` tool inside `mcp-tool-server` to allow local clients to query the 10,000+ shape library (AWS, Azure, GCP, Cisco, Kubernetes, etc.) for stencils.
